Feat: Criado as exceções para UPDATE e DELETE

+++ src/main/java/com/desafio/clients_crud/services/ClientService.java
@@ -3,11 +3,15 @@ package com.desafio.clients_crud.services;
 import com.desafio.clients_crud.dto.ClientDTO;
 import com.desafio.clients_crud.entities.Client;
 import com.desafio.clients_crud.repositories.ClientRepository;
+import com.desafio.clients_crud.services.exceptions.DatabaseException;
 import com.desafio.clients_crud.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -40,17 +44,25 @@ public class ClientService {
 
     @Transactional()
     public ClientDTO update (Long id, ClientDTO dto) {
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
 
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-
-        return new ClientDTO(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete (Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        
         repository.deleteById(id);
+
     }
